feat: align NAP/Karpenter node reservations with AKS node hardening
