+++ pkg/operator/options/options.go
@@ -66,7 +66,10 @@ type Options struct {
 	ClusterName             string  `json:"clusterName,omitempty"`
 	ClusterEndpoint         string  `json:"clusterEndpoint,omitempty"` // => APIServerName in bootstrap, except needs to be w/o https/port
 	VMMemoryOverheadPercent float64 `json:"vmMemoryOverheadPercent,omitempty"`
-	// EnableNodeHardening is a temporary preview only flag on ManagedCluster.
+	// EnableNodeHardening records the temporary ManagedCluster preview flag, but does not by itself
+	// mean node hardening is active. The bootstrappingclient mode intentionally ignores this flag
+	// because its bootstrap backend does not apply the matching kubelet configuration. Consumers
+	// must use ShouldUseNodeHardening instead of reading this field directly.
 	// Will remove this option when AKS defaults node hardening for new nodes based on Kubernetes version
 	// (expected 1.39+) and removes the corresponding field from the preview API.
 	EnableNodeHardening            bool   `json:"enableNodeHardening,omitempty"`
@@ -112,7 +115,7 @@ func (o *Options) AddFlags(fs *coreoptions.FlagSet) {
 	fs.StringVar(&o.ClusterName, "cluster-name", env.WithDefaultString("CLUSTER_NAME", ""), "[REQUIRED] The kubernetes cluster name for resource tags.")
 	fs.StringVar(&o.ClusterEndpoint, "cluster-endpoint", env.WithDefaultString("CLUSTER_ENDPOINT", ""), "[REQUIRED] The external kubernetes cluster endpoint for new nodes to connect with.")
 	fs.Float64Var(&o.VMMemoryOverheadPercent, "vm-memory-overhead-percent", utils.WithDefaultFloat64("VM_MEMORY_OVERHEAD_PERCENT", 0.075), "The VM memory overhead as a percent that will be subtracted from the total memory for all instance types.")
-	fs.BoolVar(&o.EnableNodeHardening, "enable-node-hardening", env.WithDefaultBool("ENABLE_NODE_HARDENING", false), "If set to true, Karpenter applies (kube-reserved, system-reserved, hard-eviction, soft-eviction) matching AKS node hardening, keeping scheduling simulation and rendered kubelet configuration aligned with hardened values.")
+	fs.BoolVar(&o.EnableNodeHardening, "enable-node-hardening", env.WithDefaultBool("ENABLE_NODE_HARDENING", false), "Requests AKS node hardening reservations and eviction settings where supported. This option is ignored in bootstrappingclient mode.")
 	fs.StringVar(&o.KubeletClientTLSBootstrapToken, "kubelet-bootstrap-token", env.WithDefaultString("KUBELET_BOOTSTRAP_TOKEN", ""), "[REQUIRED] The bootstrap token for new nodes to join the cluster.")
 	fs.StringVar(&o.LinuxAdminUsername, "linux-admin-username", env.WithDefaultString("LINUX_ADMIN_USERNAME", "azureuser"), "The admin username for Linux VMs.")
 	fs.StringVar(&o.SSHPublicKey, "ssh-public-key", env.WithDefaultString("SSH_PUBLIC_KEY", ""), "[REQUIRED] VM SSH public key.")
@@ -150,6 +153,13 @@ func (o *Options) AddFlags(fs *coreoptions.FlagSet) {
 // IsAKSMachineAPIMode returns true if the current provision mode creates instances via the AKS Machine API.
 func (o *Options) IsAKSMachineAPIMode() bool {
 	return o.ProvisionMode == consts.ProvisionModeAKSMachineAPI || o.ProvisionMode == consts.ProvisionModeAKSMachineAPIHeaderBatch
+}
+
+// ShouldUseNodeHardening returns the effective node hardening setting. Scriptless and AKS Machine
+// API modes apply matching kubelet configuration; bootstrappingclient does not and is excluded.
+func (o *Options) ShouldUseNodeHardening() bool {
+	return o.EnableNodeHardening &&
+		(o.ProvisionMode == consts.ProvisionModeAKSScriptless || o.IsAKSMachineAPIMode())
 }
 
 func (o *Options) GetAPIServerName() string {
