+++ pkg/providers/instancetype/instancetype_test.go
@@ -26,29 +26,6 @@ import (
 	"github.com/Azure/karpenter-provider-azure/pkg/consts"
 )
 
-func TestShouldUseNodeHardening(t *testing.T) {
-	tests := []struct {
-		name          string
-		enabled       bool
-		provisionMode string
-		want          bool
-	}{
-		{name: "disabled for scriptless", provisionMode: consts.ProvisionModeAKSScriptless},
-		{name: "enabled for scriptless", enabled: true, provisionMode: consts.ProvisionModeAKSScriptless, want: true},
-		{name: "skipped for bootstrapping client", enabled: true, provisionMode: consts.ProvisionModeBootstrappingClient},
-		{name: "enabled for machine API", enabled: true, provisionMode: consts.ProvisionModeAKSMachineAPI, want: true},
-		{name: "enabled for machine API header batch", enabled: true, provisionMode: consts.ProvisionModeAKSMachineAPIHeaderBatch, want: true},
-		{name: "skipped for unknown mode", enabled: true, provisionMode: "unknown"},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			g := NewWithT(t)
-			g.Expect(ShouldUseNodeHardening(test.enabled, test.provisionMode)).To(Equal(test.want))
-		})
-	}
-}
-
 // These cases mirror the hardened kube-reserved memory calculation in the AKS RP.
 func TestKubeReservedResourcesHardeningParity(t *testing.T) {
 	tests := []struct {
